chore(test): restructure h2, oracle, informix
