+++ src/test/java/org/schemaspy/integrationtesting/h2/H2KeywordIT.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with SchemaSpy. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.schemaspy.integrationtesting;
+package org.schemaspy.integrationtesting.h2;
 
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -29,6 +29,7 @@ import org.schemaspy.input.dbms.service.DatabaseService;
 import org.schemaspy.input.dbms.service.SqlService;
 import org.schemaspy.model.Database;
 import org.schemaspy.model.ProgressListener;
+import org.schemaspy.model.Table;
 import org.schemaspy.testing.H2MemoryRule;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
@@ -36,11 +37,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import javax.script.ScriptException;
 import java.io.File;
 import java.io.IOException;
+import java.net.URISyntaxException;
 import java.sql.SQLException;
-import java.util.List;
-import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
@@ -50,10 +51,10 @@ import static org.mockito.BDDMockito.given;
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class H2PrimaryKeySeqnoIT {
+public class H2KeywordIT {
 
     @ClassRule
-    public static H2MemoryRule h2MemoryRule = new H2MemoryRule("pkorder").addSqlScript("src/test/resources/integrationTesting/h2/dbScripts/pkordering.sql");
+    public static H2MemoryRule h2MemoryRule = new H2MemoryRule("h2keyword").addSqlScript("src/test/resources/integrationTesting/h2/dbScripts/keyword_in_table.sql");
 
     @Autowired
     private SqlService sqlService;
@@ -73,7 +74,7 @@ public class H2PrimaryKeySeqnoIT {
     private static Database database;
 
     @Before
-    public synchronized void createDatabaseRepresentation() throws SQLException, IOException {
+    public synchronized void createDatabaseRepresentation() throws SQLException, IOException, ScriptException, URISyntaxException {
         if (database == null) {
             doCreateDatabaseRepresentation();
         }
@@ -82,17 +83,17 @@ public class H2PrimaryKeySeqnoIT {
     private void doCreateDatabaseRepresentation() throws SQLException, IOException {
         String[] args = {
                 "-t", "src/test/resources/integrationTesting/dbTypes/h2memory",
-                "-db", "pkorder",
-                "-s", "pkorder",
-                "-o", "target/integrationtesting/h2pkorder",
+                "-db", "h2keyword",
+                "-s", "h2keyword",
+                "-o", "target/integrationtesting/h2keyword",
                 "-u", "sa"
         };
-        given(arguments.getOutputDirectory()).willReturn(new File("target/integrationtesting/h2pkorder"));
+        given(arguments.getOutputDirectory()).willReturn(new File("target/integrationtesting/h2keyword"));
         given(arguments.getDatabaseType()).willReturn("src/test/resources/integrationTesting/dbTypes/h2memory");
         given(arguments.getUser()).willReturn("sa");
         given(arguments.getCatalog()).willReturn(h2MemoryRule.getConnection().getCatalog());
         given(arguments.getSchema()).willReturn(h2MemoryRule.getConnection().getSchema());
-        given(arguments.getDatabaseName()).willReturn("pkorder");
+        given(arguments.getDatabaseName()).willReturn("h2keyword");
         Config config = new Config(args);
         sqlService.connect(config);
         Database database = new Database(
@@ -102,12 +103,17 @@ public class H2PrimaryKeySeqnoIT {
                 arguments.getSchema()
         );
         databaseService.gatherSchemaDetails(config, database, null, progressListener);
-        this.database = database;
+        H2KeywordIT.database = database;
     }
 
     @Test
-    public void primaryKeysShouldBeInCorrectOrder() {
-        List<String> pkcolumnNames = database.getTablesMap().get("TABLE1").getPrimaryColumns().stream().map(t -> t.getName()).collect(Collectors.toList());
-        assertThat(pkcolumnNames).containsExactly("ZZ", "AA", "BB");
+    public void databaseShouldExist() {
+        assertThat(database).isNotNull();
+        assertThat(database.getName()).isEqualToIgnoringCase("h2keyword");
+    }
+
+    @Test
+    public void tableWithKeyWordShouldExist() {
+        assertThat(database.getTables()).extracting(Table::getName).contains("DISTINCT");
     }
 }
