+++ src/test/resources/integrationTesting/h2/dbScripts/2tables1view.sql
