+++ src/test/resources/integrationTesting/h2/dbScripts/spaces_in_schema_and_table.sql
