+++ src/test/resources/integrationTesting/informix/dbScripts/informixcc.sql
