+++ src/test/resources/integrationTesting/oracle/dbScripts/spaces_in_table_names.sql
